requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
django-countries==3.3
django-durationfield==0.5.1
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6
static3==0.6.1
Unidecode==0.4.17
wheel==0.24.0
